- Implemented API server - Implemented multi-threading for Discovery and Messaging servers - Added javadoc - Optimized project package structure

+++ alpha4/src/main/java/core/model/Response.java
@@ -4,6 +4,9 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.util.Map;
 
-public record Response(Status status, Map<Integer, Message> messages, String message,
+/**
+ * Messaging response record.
+ */
+public record Response(Status status, Map<Long, Message> messages, String message,
         @JsonProperty("peer_id") String peerId) {
 }
